+++ docs/guide/source/changelog.rst
@@ -18,9 +18,16 @@ update of MuJoCo alone can increase the major version.
 
 Unreleased
 ================================
-- :docs-rs:`~mujoco_rs::viewer::<struct>MjViewer`:
+- Option to automatically pull MuJoCo.
+- pkg-config support for Linux and MacOS. Note that this is not officially supported by MuJoCo.
+- Changes to :docs-rs:`~mujoco_rs::viewer::<struct>MjViewer`:
 
   - Added an optional (enabled by default) user interface, made with `egui <https://github.com/emilk/egui>`_.
+
+- New Cargo features:
+
+  - ``auto-download-mujoco``: will automatically download and setup MuJoCo (Windows and Linux).
+  - ``use-rpath``: will embed the path to the MuJoCo C library into the RPATH of the binary (Linux and MacOS).
 
 2.0.1 (MuJoCo 3.3.7)
 ================================